caculate label data

+++ segmentation/dataset/LungDataset.py
@@ -147,8 +147,8 @@ class LungDataset(Dataset):
         for i in range(batch_size):
             
             img = self.de_normalize(list_imgs[i])
-            img = int(conver_numpy_image(img))
-            mask = int(conver_numpy_image(list_masks[i]))
+            img = conver_numpy_image(img).astype(np.uint8)
+            mask = conver_numpy_image(list_masks[i]).astype(np.uint8)
             # import pdb; pdb.set_trace()
             # img = int((img +1)*255)
             ct = contour(img, mask)
